:sparkles: Commit Before Merge
