+++ src/Router.js
@@ -1,12 +1,12 @@
 import { Route, Routes } from 'react-router-dom';
-import { Navigation } from 'components/molecules';
+import { Header } from 'components/molecules';
 
 import { MakeForm, MakePost, Login, SignIn } from 'pages';
 
 const Router = () => {
   return (
     <>
-      <Navigation />
+      <Header />
       <Routes>
         <Route path="/makeform" element={<MakeForm />} />
         <Route path="/signin" element={<SignIn />} />
