+++ src/components/molecules/index.js
@@ -1,1 +1,1 @@
-export * from './Navigation';
+export * from './Header';
